#381: move dataframehelper module to external resource
