+++ src/test/python/dataframehelper/stage/modules.py
@@ -1,0 +1,30 @@
+#
+# This file is licensed under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from smv import *
+from smv.dqm import *
+from pyspark.sql.functions import col, lit
+
+
+class D1(SmvPyCsvStringData):
+    def schemaStr(self):
+        return "k:String;v:Integer"
+    def dataStr(self):
+        return "a,1;b,2"
+
+class T(SmvPyCsvFile):
+    @classmethod
+    def path(cls):
+        return "./target/python-test-export-csv.csv"
+    def csvAttr(self):
+        return self.defaultCsvWithHeader()
