+++ src/test/python/smvframework/stage/modules.py
@@ -11,7 +11,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# from smv import SmvApp, SmvModule, SmvOutput
 from smv import *
 from smv.dqm import *
 from pyspark.sql.functions import col, lit
