+++ src/test/python/testDataFrameHelper.py
@@ -12,7 +12,6 @@
 # limitations under the License.
 
 import unittest
-import sys
 
 from test_support.smvbasetest import SmvBaseTest
 from smv import SmvApp, SmvPyCsvFile
@@ -23,18 +22,12 @@ import pyspark
 from pyspark.context import SparkContext
 from pyspark.sql import SQLContext, HiveContext
 from pyspark.sql.functions import col, struct
-
-class T(SmvPyCsvFile):
-    @classmethod
-    def path(cls):
-        return "./target/python-test-export-csv.csv"
-    def csvAttr(self):
-        return self.defaultCsvWithHeader()
+from dataframehelper.stage.modules import D1, T
 
 class DfHelperTest(SmvBaseTest):
     @classmethod
     def smvAppInitArgs(cls):
-        return ['--smv-props', 'smv.stages=testDataFrameHelper']
+        return ['--smv-props', 'smv.stages=dataframehelper.stage']
 
     def test_smvGroupBy(self):
         return "TODO implement"
@@ -128,13 +121,11 @@ class DfHelperTest(SmvBaseTest):
         self.should_be_same(expect, res)
 
     def test_smvExportCsv(self):
-        global oldmod
-        oldmod = sys.modules["testDataFrameHelper"]
-
-        df = self.createDF("k:String;v:Integer", "a,1;b,2")
+        fqn = D1.fqn()
+        df = self.df(fqn)
         df.smvExportCsv(T.path())
 
-        res = self.smvApp.runModule("mod:" + self.__module__ + ".T")
+        res = self.smvApp.runModule("mod:" + "dataframehelper.stage.modules.T")
         self.should_be_same(df, res)
 
     def test_smvJoinByKey(self):
